+++ lib/notifier/index.js
@@ -1,66 +1,13 @@
-const fetch = require('node-fetch')
-
-function doGet(url, data) {
-  const params = new URLSearchParams(data)
-
-  return fetch(`${url}?${params.toString()}`, {
-    timeout: 10000
-  }).then((res) => res.json())
-}
-
-function doPost(url, data, type = 'form') {
-  let cType, body
-
-  if (type == 'json') {
-    cType = 'application/json'
-    body = JSON.stringify(data)
-  } else {
-    const params = new URLSearchParams(data)
-
-    cType = 'application/x-www-form-urlencoded'
-    body = params.toString()
-  }
-
-  return fetch(url, {
-    method: 'POST',
-    body: body,
-    headers: {
-      'Content-Type': cType
-    },
-    timeout: 10000
-  }).then((res) => res.json())
-}
-
-/**
- * 获取企业微信 accessToken
- * https://work.weixin.qq.com/api/doc/90000/90135/91039
- *
- * @param  {[type]} corpId     企业 id
- * @param  {[type]} corpSecret 企业应用密匙
- * @return {Promise<Object>}   结果
- */
-async function getQywxAccessToken(corpId, corpSecret) {
-  const res = await doGet('https://qyapi.weixin.qq.com/cgi-bin/gettoken', {
-    corpid: corpId,
-    corpsecret: corpSecret
-  })
-
-  if (res.errcode != 0) {
-    throw new Error('AccessToken 获取失败：' + res.errmsg)
-  }
-
-  // 提前十分钟过期
-  const expiresTime = Date.now() + res.expires_in * 1000 - 10 * 60 * 1000
-
-  return {
-    token: res.access_token,
-    expires: expiresTime
-  }
-}
+const sendBark = require('./vendor/bark')
+const sendLark = require('./vendor/lark')
+const sendServerChan = require('./vendor/server-chan')
+const sendTelegram = require('./vendor/telegram')
+const { sendWorkWechat, getQywxAccessToken } = require('./vendor/work-wechat')
 
 class Notifier {
   constructor(options) {
     this.barkKey = options.barkKey
+    this.larkWebhook = options.larkWebhook
     this.serverChanToken = options.serverChanToken
     this.telegram = options.telegram
     this.workWechat = null
@@ -112,17 +59,28 @@ class Notifier {
       }
     }
 
-    const url = `https://api.day.app/${barkKey}/`
-    const path = [title, content].map(encodeURIComponent).join('/')
-    let data
+    return sendBark({ title, content, link, pushKey: barkKey })
+  }
 
-    if (link) {
-      data = { url: link }
+  /**
+   * 飞书通知
+   *
+   * @param  {String} title   标题
+   * @param  {String} content 内容
+   * @param  {Object} options  配置
+   * @return {Promise<String>}    推送结果
+   */
+  async sendLark(title = '', content = '', { webhook }) {
+    const larkWebhook = webhook || this.larkWebhook
+
+    if (!larkWebhook) {
+      return {
+        success: false,
+        msg: `飞书推送失败: 请设置 larkWebhook`
+      }
     }
 
-    return doGet(url + path, data)
-      .then((res) => ({ success: true, msg: 'Bark 推送成功' }))
-      .catch((e) => ({ success: false, msg: `Bark 推送失败: ${e}` }))
+    return sendLark({ title, content, webhook: larkWebhook })
   }
 
   /**
@@ -135,22 +93,22 @@ class Notifier {
    * @return {Promise<String>}  推送结果
    */
   async sendTelegram(title = '', content = '', { uid = '' }) {
-    if (!this.telegram || !this.telegram.botToken) {
+    const botToken = this.telegram && this.telegram.botToken
+    const user = uid || this.telegram.userId
+
+    if (!botToken) {
       return {
         success: false,
         msg: `Telegram 推送失败: 请设置 bot token`
       }
     }
 
-    const url = `https://api.telegram.org/bot${this.telegram.botToken}/sendMessage`
-    const data = {
-      chat_id: uid || this.telegram.userId,
-      text: `【${title}】\n${content}`
-    }
-
-    return doPost(url, data)
-      .then((res) => ({ success: true, msg: 'Telegram 推送成功' }))
-      .catch((e) => ({ success: false, msg: `Telegram 推送失败: ${e}` }))
+    return sendTelegram({
+      botToken,
+      content,
+      title,
+      user
+    })
   }
 
   /**
@@ -162,13 +120,7 @@ class Notifier {
    * @return {Promise<String>}  推送结果
    */
   async sendServerChan(title = '', content = '') {
-    const api = this.serverChanToken.includes('SCT') ? 'sctapi' : 'sc'
-    const url = `https://${api}.ftqq.com/${this.serverChanToken}.send`
-    const data = { title, desp: content }
-
-    return doPost(url, data)
-      .then((res) => ({ success: true, msg: 'Server 酱推送成功' }))
-      .catch((e) => ({ success: false, msg: `Server 酱推送失败: ${e}` }))
+    return sendServerChan({ title, content, token: this.serverChanToken })
   }
 
   /**
@@ -189,30 +141,14 @@ class Notifier {
     }
 
     const { agentId, toUser } = this.workWechat
+    const user = uid || toUser || '@all'
 
-    try {
-      const accessToken = await this._getQywxAccessToken()
-      const res = await doPost(
-        `https://qyapi.weixin.qq.com/cgi-bin/message/send?access_token=${accessToken}`,
-        {
-          touser: uid || toUser || '@all',
-          msgtype: 'text',
-          agentid: agentId,
-          text: {
-            content: `【${title}】\n${content}`
-          }
-        },
-        'json'
-      )
-
-      if (res.errcode != 0) {
-        throw res.errmsg
-      }
-    } catch (e) {
-      return { success: false, msg: `企业微信推送失败: ${e}` }
-    }
-
-    return { success: true, msg: '企业微信推送成功' }
+    return sendWorkWechat({
+      agentId,
+      content,
+      title,
+      user
+    })
   }
 
   /**
@@ -228,6 +164,10 @@ class Notifier {
 
     if (this.barkKey) {
       result.push(this.sendBark(title, content, options))
+    }
+
+    if (this.larkWebhook) {
+      result.push(this.sendLark(title, content, options))
     }
 
     if (this.telegram && this.telegram.botToken) {
